+++ semcore/counter/src/index.d.ts
@@ -5,7 +5,7 @@ import { BoxProps } from '@semcore/flex-box';
 export interface ICounterProps extends CounterProps, UnknownProperties {}
 export type CounterProps = BoxProps & {
   /** Counter theme or custom color */
-  theme?: 'warning' | 'danger' | string;
+  theme?: 'warning' | 'danger' | 'info' | string;
 
   /** Counter size
    * @default m */
@@ -21,9 +21,22 @@ export type AnimatedNumberBaseProps = {
   easing?: (t: number) => number;
   /** Stringify number, receives a fraction value */
   formatValue?: (value: number) => string;
+  /**
+   * Duration time in ms
+   * @default 200
+   * */
   duration?: number;
+  /**
+   * Delay before animation in ms
+   * @default 0
+   * */
   delay?: number;
+  /**
+   * The value from which to start the animation
+   * @default 0
+   * */
   initValue?: number;
+  /** The value by which to end the animation */
   value: number;
 };