+++ website/docs/components/dot/examples/example_of_dot_animation.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import NotificationM from 'intergalactic/icon/Notification/m';
 import Button from 'intergalactic/button';
 import Dot from 'intergalactic/dot';
+import { Hint } from 'intergalactic/tooltip';
 
 const Demo = () => {
   const [dotVisible, setDotVisible] = React.useState(true);
@@ -14,10 +15,12 @@ const Demo = () => {
   const handleClick = React.useCallback(() => setDotVisible(false), []);
 
   return (
-    <Button onClick={handleClick} aria-label='Read notifications'>
-      <NotificationM />
-      <Dot up hidden={!dotVisible} size='l' aria-label='You have notifications to read' />
-    </Button>
+    <Hint tag={Button} onClick={handleClick} title='Notifications'>
+      <Button.Addon>
+        <NotificationM />
+        <Dot up hidden={!dotVisible} size='l' />
+      </Button.Addon>
+    </Hint>
   );
 };