+++ semcore/data-table/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.40.4] - 2024-06-14
+
+### Added
+
+- Possibility to pass `scrollContainerRef` property to `Datatable.Body` and set it to a real table body container - `Scroll.Containter`.
+
+## [4.40.3] - 2024-06-13
+
+### Fixed
+
+- Incorrect display of `ScrollArea.Bar` after reducing the parent size.
+
 ## [4.40.2] - 2024-06-12
 
 ### Changed