+++ semcore/dropdown/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.32.2] - 2024-06-14
+
+### Fixed
+
+- Dropdown with any `interaction` except `none` should be opened by pressing `Enter` or `Space`.
+
+## [4.32.1] - 2024-06-14
+
+### Fixed
+
+- Dropdown with `interaction='none'` shouldn't be opened by pressing `Enter` or `Space`.
+
+## [4.32.0] - 2024-06-13
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/utils` [4.28.2 ~> 4.29.0], `@semcore/core` [2.25.2 ~> 2.26.0]).
+
 ## [4.31.7] - 2024-06-12
 
 ### Changed