+++ semcore/illustration/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.28.0] - 2024-06-13
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/flex-box` [5.27.2 ~> 5.28.0], `@semcore/core` [2.25.2 ~> 2.26.0]).
+
 ## [2.27.2] - 2024-06-12
 
 ### Changed