+++ semcore/notice/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.35.0] - 2024-06-19
+
+### Changed
+
+- Notice with `warning` and `danger` themes now has aria-label "Critical Notification" (other teams has "Notification" as before).
+
+### Added
+
+- `muted` theme support.
+
+## [5.34.3] - 2024-06-13
+
+### Fixed
+
+- Now `string` is not an allowed value for theme prop.
+
 ## [5.34.2] - 2024-06-12
 
 ### Changed